backend/requirements.txt
Django==3.2.7
pytz==2022.7.1
sqlparse==0.4.3
Pillow==9.4.0
djangorestframework==3.14.0
sorl-thumbnail==12.9.0
drf-extra-fields==3.4.1
python-dotenv==0.21.1
psycopg2-binary==2.9.5
